requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
groq==0.13.0
google-cloud-texttospeech==2.21.0
gunicorn==23.0.0